+++ SelfHost/Handlers/Cookies.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using Simple.Web;
 using Simple.Web.Behaviors;
 using Simple.Web.Http;
+using Simple.Web.Owin;
 
 namespace SelfHost
 {
@@ -22,24 +23,6 @@ namespace SelfHost
 
 		public IDictionary<string, ICookie> RequestCookies { get; set; }
     }
-
-	public class SimpleCookie : ICookie {
-		public SimpleCookie(string name, string value) {
-			Name = name;
-			Value = value;
-		}
-
-		public string Name { get; set; }
-		public bool Secure { get; set; } 
-		public bool HttpOnly{get;set;} 
-		public DateTime Expires{get;set;} 
-		public string Value{get;set;} 
-		public IDictionary<string, string> Values{get;set;} 
-		public string this[string key] {
-			get { return Values[key]; }
-			set { Values[key] = value; }
-		}
-	}
 
 	[UriTemplate("/addcookie")]
 	public class AddCookie : IPost, IMayRedirect, ISetCookies, IInput<CookieText> {
