+++ SelfHost/Views/md5Upload.cshtml
@@ -7,7 +7,7 @@
 <body>
     <h1> Upload a file</h1>
     <div>
-        <form action="/md5" method="POST" ><!--enctype="multipart/form-data">-->
+        <form action="/md5" method="POST" enctype="multipart/form-data">
         <fieldset>
             <div>
                 <label for="datafile">File:</label>
