+++ Simple.Web.Owin/RequestWrapper.cs
@@ -119,11 +119,22 @@ namespace Simple.Web.Owin
 			return boundsLine.SubstringAfterLast('=');
 		}
 
-		Dictionary<string, ICookie> cookieCache = new Dictionary<string, ICookie>();
+		Dictionary<string, ICookie> cookies;
 		public IDictionary<string, ICookie> Cookies
 		{
 			get {
-				return cookieCache; //TODO
+				if (cookies == null) {
+					cookies = new Dictionary<string, ICookie>();
+					if (!headers.ContainsKey("Cookie")) return cookies;
+					var cookieStrings = headers["Cookie"].First().Split(new[] { ';' }, StringSplitOptions.RemoveEmptyEntries);
+					foreach (var cookie in cookieStrings) {
+						var p = cookie.Split('=');
+						var k = p[0];
+						var v = p.Length > 1 ? p[1] : p[0];
+						cookies.Add(k, new SimpleCookie(k, v));
+					}
+				}
+				return cookies;
 			}
 		}
 	}
