+++ Simple.Web.Owin/Simple.Web.Owin.csproj
@@ -54,6 +54,7 @@
     <Compile Include="RequestWrapper.cs" />
     <Compile Include="ResponseWrapper.cs" />
     <Compile Include="SelfHostingApp.cs" />
+    <Compile Include="SimpleCookie.cs" />
     <Compile Include="StringExtensions.cs" />
     <Compile Include="TaskEx.cs" />
     <Compile Include="wtf.cs" />
